fix(home-mentions): 修复 UI Type Checks / Build Smoke 失败 — runAgent 类型扩展 + route.ts 导出面收敛;
- 前因：feat(home-mentions) 在 home-body.tsx 调用 ``agent.runAgent({ runId, forwardedProps })``，但 ``HomeBodyAgent.runAgent`` 类型未同步扩展，TS2353 失败；同时 ``app/api/agui/route.ts`` 直接 ``export`` 工具函数 ``buildStateDeltaFromForwardedProps``，违反 Next.js App Router 对 route 模块「仅可导出 HTTP handler / 配置符号」的契约，触发 OmitWithTag 路由类型校验失败。两者共同阻塞 ui-quality / UI Type Checks 与 UI Build Smoke 两条 CI；
- 修复一：``HomeBodyAgent.runAgent`` 入参类型新增可选 ``forwardedProps?: Record<string, unknown>``，与同 type 已声明的 ``forwardedProps`` 实例属性同型，并对齐 ``@ag-ui/client`` 的 ``RunAgentParameters`` 语义子集；旧调用方仅传 ``{ runId }`` 保持兼容；
- 修复二：将 ``buildStateDeltaFromForwardedProps`` 与共享的 ``UUID_RE`` / ``PREFERRED_SUBAGENT_MAX_LEN`` / ``CORPUS_IDS_MAX_LEN`` 常量及 ``sanitizeUuidList`` 私有辅助下沉到 sibling 私有模块 ``apps/negentropy-ui/app/api/agui/_state-delta.ts``，与 ``sessions/_request.ts`` ``_response.ts`` 形成同构（项目约定的 ``_`` 前缀私有模块）；``route.ts`` 改为从该模块 import，导出面收敛为合规的 ``POST`` 与未来潜在的 HTTP/配置符号；
- 同步：``tests/unit/api/agui-route-state.test.ts`` 切换 import 到新模块，保持单一事实源；
- 验证：``pnpm typecheck`` 与 ``pnpm build`` 均通过；91 个测试文件 / 709 个单元测试全部绿灯。
